Move get_vars_for_user().
