+++ evictionfree/hardship_declaration.py
@@ -1,5 +1,7 @@
+from datetime import date
 from typing import List, Optional
 from pathlib import Path
+from users.models import JustfixUser
 from pydantic import BaseModel
 
 from .overlay_pdf import Text, Checkbox, Page, Document
@@ -92,3 +94,23 @@ def fill_hardship_pdf(v: HardshipDeclarationVariables, locale: str) -> bytes:
     assert path.exists()
     overlay = Document(pages=get_pages(v, locale))
     return overlay.overlay_atop(path).getvalue()
+
+
+def get_vars_for_user(user: JustfixUser) -> Optional[HardshipDeclarationVariables]:
+    if not (
+        user.is_authenticated
+        and hasattr(user, "hardship_declaration_details")
+        and hasattr(user, "onboarding_info")
+    ):
+        return None
+    hdd = user.hardship_declaration_details
+    onb = user.onboarding_info
+    return HardshipDeclarationVariables(
+        index_number=hdd.index_number or None,
+        county_and_court=None,  # TODO: Fill this out!
+        address=", ".join(onb.address_lines_for_mailing),
+        has_financial_hardship=hdd.has_financial_hardship,
+        has_health_risk=hdd.has_health_risk,
+        name=user.full_name,
+        date=date.today().strftime("%m/%d/%Y"),
+    )
