+++ evictionfree/tests/conftest.py
@@ -3,9 +3,9 @@ import pytest
 
 @pytest.fixture
 def fake_fill_hardship_pdf(monkeypatch):
-    from evictionfree import views
+    from evictionfree import hardship_declaration
 
     def fake(v, locale):
         return bytes(b"blah")
 
-    monkeypatch.setattr(views, "fill_hardship_pdf", fake)
+    monkeypatch.setattr(hardship_declaration, "fill_hardship_pdf", fake)
